solr_mint_package/requirements.txt
xrpl-py==2.6.0
PyYAML==6.0.2
Pillow==10.4.0
qrcode==7.4.2
python-dotenv==1.0.1
requests==2.32.3
Flask==2.3.2
export XUMM_API_KEY="your_key_here"
export XUMM_API_SECRET="your_secret_here"